Delete n Edit Authorization

+++ backend/routes/posts.js
@@ -65,7 +65,8 @@ router.post('', checkJWT, multer({storage}).single('image'), (req, res, next) =>
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + '/images/' + req.file.filename
+    imagePath: url + '/images/' + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then((createdPost) => {
     res.status(201).json({
@@ -90,14 +91,23 @@ router.put('/:id',checkJWT, multer({storage}).single('image'), (req, res, next) 
     content: req.body.content,
     imagePath: imagePath
   });
-  Post.updateOne({_id: req.params.id}, post).then(updatedPost => {
-    res.status(200).json({ message: 'Updated Successfully!' });
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId }, post).then(updatedPost => {
+    if (updatedPost.nModified > 0) {
+      res.status(200).json({ message: 'Updated Successfully!' });
+    } else {
+      res.status(401).json({ message: 'Not Authorization' });
+    }
+
   })
 });
 
 router.delete('/:id', checkJWT, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then((result) => {
-    res.status(200).json({ message: 'post deleted successfully!' });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then((result) => {
+    if (updatedPost.n > 0) {
+      res.status(200).json({ message: 'post deleted successfully!' });
+    } else {
+      res.status(401).json({ message: 'Not Authorization' });
+    }
   })
   .catch((reason) => {
     res.status(500).json({ message: 'Something went wrong!' });
